backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic-settings==2.2.1
httpx==0.27.0
PyJWT==2.8.0
sqlalchemy==2.0.27
aiosqlite==0.20.0
python-dotenv==1.0.1
